+++ voip_sip_webrtc_voice/__manifest__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 {
     'name': "Voip Communication - Voice Synthesis",
-    'version': "1.0.1",
+    'version': "1.0.2",
     'author': "Sythil Tech",
     'category': "Tools",
     'support': "steven@sythiltech.com.au",
@@ -10,7 +10,10 @@
     'data': [
         'views/voip_settings_views.xml',
         'views/voip_call_template_views.xml',
+        'views/voip_voice_message_views.xml',
+        'views/voip_account_action_views.xml',
         'data/voip.voice.csv',
+        'data/voip.account.action.type.csv',
         'data/voip_settings.xml',
     ],
     'demo': [],