+++ voip_sip_webrtc_voice/models/__init__.py
@@ -1,1 +1,1 @@
-import voip_voice, voip_call_template, voip_settings
+import voip_voice, voip_call_template, voip_settings, voip_voice_message, voip_account_action