+++ voip_sip_webrtc_voice/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.2
+======
+* Voice synth on incoming call
+
 v1.0.1
 ======
 * Bing TTS